requirements.txt
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0
unidecode>=1.3.0
sqlalchemy>=2.0.0
openpyxl>=3.0.0